+++ bootstrap/src/main/java/onlydust/com/marketplace/api/bootstrap/configuration/BackofficeConfiguration.java
@@ -6,6 +6,7 @@ import onlydust.com.marketplace.api.domain.port.output.BackofficeStoragePort;
 import onlydust.com.marketplace.api.domain.service.BackofficeService;
 import onlydust.com.marketplace.api.rest.api.adapter.BackofficeCurrencyManagementRestApi;
 import onlydust.com.marketplace.api.rest.api.adapter.BackofficeRestApi;
+import onlydust.com.marketplace.api.rest.api.adapter.BackofficeSponsorManagementRestApi;
 import onlydust.com.marketplace.api.rest.api.adapter.authentication.api_key.ApiKeyAuthenticationService;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -24,6 +25,11 @@ public class BackofficeConfiguration {
     @Bean
     public BackofficeCurrencyManagementRestApi backofficeCurrencyManagementRestApi(final CurrencyFacadePort currencyFacadePort) {
         return new BackofficeCurrencyManagementRestApi(currencyFacadePort);
+    }
+
+    @Bean
+    public BackofficeSponsorManagementRestApi backofficeSponsorManagementRestApi(final BackofficeFacadePort backofficeFacadePort) {
+        return new BackofficeSponsorManagementRestApi(backofficeFacadePort);
     }
 
     @Bean
